lms_backend/requirements.txt
# ============================================================================
# IONIA LMS BACKEND - PRODUCTION-GRADE TECH STACK (2025)
# ============================================================================


# ============================================================================
# CORE FRAMEWORK
# ============================================================================
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
python-multipart==0.0.12


# ============================================================================
# SUPABASE & DATABASE
# ============================================================================
supabase==2.22.2
# Remove explicit postgrest - it's managed by supabase automatically
sqlalchemy[asyncio]==2.0.36
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9


# ============================================================================
# DATA VALIDATION
# ============================================================================
pydantic==2.10.1
pydantic-settings==2.6.0
email-validator==2.1.0


# ============================================================================
# AI & LLM STACK
# ============================================================================
litellm==1.79.0
praisonaiagents==0.0.162
google-cloud-vision==3.8.0
qdrant-client==1.12.1
sentence-transformers==3.3.1
torch==2.2.0
pillow==11.0.0


# ============================================================================
# BACKGROUND TASKS & CACHING
# ============================================================================
celery[redis]==5.4.0
redis[asyncio]==5.0.0
flower==2.0.1
hiredis==2.3.2


# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
PyJWT==2.8.0


# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.0
python-dateutil==2.8.2
pytz==2024.1


# ============================================================================
# EXPORT & REPORTING
# ============================================================================
reportlab==4.0.9
openpyxl==3.1.2
pandas==2.1.4


# ============================================================================
# LOGGING & MONITORING
# ============================================================================
loguru==0.7.0
python-json-logger==2.0.7


# ============================================================================
# TESTING
# ============================================================================
pytest==8.2.2
pytest-asyncio==0.26.0
pytest-cov==5.0.0
faker==25.0.0


# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================
black==24.8.0
ruff==0.6.0
mypy==1.11.0
pre-commit==3.8.0
